requirements.txt
torch
open_clip_torch
numpy
pandas
scikit-learn
tqdm
loguru
